Remove `active_thread`
